fixed truncation bug

+++ main.cpp
@@ -50,7 +50,7 @@ public:
 
         // Block N1
         for (int i = 0; i < N_; ++i) {
-            int N_one_coeff = 2.0 * (2.0 * (i + 1) + 1.0) * (1.0/((i + 1) * ((i + 1) + 1.0)));
+            double N_one_coeff = 2.0 * (2.0 * (i + 1) + 1.0) * (1.0/((i + 1) * ((i + 1) + 1.0)));
             add(A, i, 5 * i, -N_one_coeff * (i + 1) * ((i + 1) + 2.0) * lambda);
             add(A, i, 5 * i + 1, -N_one_coeff * ((i + 1) * (i + 1) - 1.0) * lambda);
             add(A, i, 5 * i + 2, -N_one_coeff * ((i + 1) * (i + 1) - 1.0) * lambda);
